wrap art resampler into a component
